kvserver: Move MultiSSTWriter to storage package
This pr moves MultiSSTWriter to the storage package. This is done so
that its logic can be reused in multiple different places such as
backups.

Release note: None
Informs: #67824
